+++ packages/ts-components/src/components/olympics/medal-table/OlympicsMedalTable.stories.tsx
@@ -2,34 +2,24 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 
 import { OlympicsMedalTable } from './OlympicsMedalTable';
-import { HelmetProvider } from 'react-helmet-async';
 import { ArticleHarness } from '../../../fixtures/article-harness/ArticleHarness';
-import { text, select } from '@storybook/addon-knobs';
+import { boolean, text } from '@storybook/addon-knobs';
+import { config } from '../OlympicsKeys';
 
 storiesOf('Typescript Component/Olympics', module)
   .addDecorator((storyFn: () => React.ReactNode) => (
-    <HelmetProvider context={{}}>
-      <ArticleHarness>{storyFn()}</ArticleHarness>
-    </HelmetProvider>
+    <ArticleHarness>{storyFn()}</ArticleHarness>
   ))
 
   .add('Medal Table', () => {
-    const endpoint = select(
-      'Endpoint',
-      {
-        staging: 'https://olympics-embed-staging.pamedia.io',
-        prod: 'https://olympics-embed.pamedia.io'
-      },
-      'https://olympics-embed-staging.pamedia.io'
-    );
-    const authToken = text('Auth Token', '6i3DuEwbVhr2Fht6');
-    const gamesCode = text('Games Code', 'OG2020-TR2');
+    const inArticle = boolean('Is In Article', true);
     const highlighted = text('Highlighted Country', 'GBR');
 
     return (
       <OlympicsMedalTable
-        keys={{ endpoint, authToken, gamesCode }}
+        keys={config.prod}
         highlighted={highlighted}
+        inArticle={inArticle}
       />
     );
   });