+++ packages/ts-components/src/components/olympics/schedule/OlympicsSchedule.tsx
@@ -1,27 +1,50 @@
-import React, { FC, useState } from 'react';
-import { Helmet } from 'react-helmet-async';
-import { Container } from './styles';
-import { colours } from '@times-components/styleguide';
+import React, { FC, useEffect, useState } from 'react';
 
-import { OlympicsKeys } from '../types';
-import { Button } from '../medal-table/styles';
+import { Container } from './styles';
+import { HeadingContainer, Label, Button, Heading } from '../shared-styles';
+
+import { injectScript } from '../../../helpers/widgets/inject-script';
+import { OlympicsKeys, config } from '../OlympicsKeys';
 
 export const OlympicsSchedule: FC<{
-  keys: OlympicsKeys;
-  sectionColor?: string;
+  keys?: OlympicsKeys;
+  inArticle?: boolean;
 }> = ({
-  keys: { endpoint, authToken, gamesCode },
-  sectionColor = colours.section.sport
+  keys: { endpoint, authToken, gamesCode } = config.prod,
+  inArticle = true
 }) => {
+  useEffect(() => {
+    injectScript(`${endpoint}/static/schedule.js`);
+  }, []);
+
+  useEffect(() => {
+    window.addEventListener(
+      'wheel',
+      event => {
+        if (
+          event
+            .composedPath()
+            // @ts-ignore
+            .includes(document.querySelector('.pa_UnitListView_ctr'))
+        ) {
+          event.stopImmediatePropagation();
+        }
+      },
+      true
+    );
+  }, []);
+
   const [showAll, setShowAll] = useState(false);
   const handleShowAll = () => {
     setShowAll(!showAll);
   };
+
   return (
-    <Container sectionColour={sectionColor} showAll={showAll}>
-      <Helmet>
-        <script src={`${endpoint}/static/schedule.js`} charSet="UTF-8" defer />
-      </Helmet>
+    <Container inArticle={inArticle} showAll={showAll}>
+      <HeadingContainer>
+        <Label>Tokyo 2020</Label>
+        <Heading>Event Schedule</Heading>
+      </HeadingContainer>
       <div
         className="pa-schedule"
         data-auth-token={authToken}